api/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# HTTP client
httpx>=0.26.0

# File handling
aiofiles>=23.2.1

# Environment
python-dotenv>=1.0.0

# CORS
# (built into FastAPI)
